+++ gamestatus/board/theboard.py
@@ -1,0 +1,3 @@
+cols_list = ['  1  ', ' 2  ', ' 3  ', ' 4  ', ' 5  ', ' 6  ', ' 7']
+board_divider = ["+---" * 7, "+"]
+board = [["|   "] * 7 for i in range(6)]
